Move persistable to postgres module
